+++ data-transfer-object/src/main/java/org/grnet/status/dtos/downtime/DailyDowntimeResponse.java
@@ -1,0 +1,15 @@
+package org.grnet.status.dtos.downtime;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class DailyDowntimeResponse {
+
+    private String date;
+
+    private List<DailyDowntimeEndpointResponse> endpoints;
+}